requirements.txt
fastapi
uvicorn
rdkit-pypi>=2022.09.4
HOSE_code_generator @ git+https://github.com/Ratsemaat/HOSE_code_generator
jpype1==1.4.1
pystow>=0.4.9